+++ programs/drift/src/state/user.rs
@@ -129,7 +129,8 @@ pub struct User {
     pub open_auctions: u8,
     /// Whether or not user has open order with auction
     pub has_open_auction: bool,
-    pub padding1: [u8; 5],
+    pub margin_mode: MarginMode,
+    pub padding1: [u8; 4],
     pub last_fuel_bonus_update_ts: u32,
     pub padding: [u8; 12],
 }
@@ -433,6 +434,10 @@ impl User {
         }
 
         false
+    }
+
+    pub fn is_high_leverage_mode(&self) -> bool {
+        self.margin_mode == MarginMode::HighLeverage
     }
 
     pub fn get_fuel_bonus_numerator(&self, now: i64) -> DriftResult<i64> {
@@ -1905,4 +1910,11 @@ pub struct ReferrerName {
 
 impl Size for ReferrerName {
     const SIZE: usize = 136;
+}
+
+#[derive(Clone, Copy, BorshSerialize, BorshDeserialize, PartialEq, Debug, Eq, Default)]
+pub enum MarginMode {
+    #[default]
+    Default,
+    HighLeverage,
 }