+++ sdk/src/user.ts
@@ -676,7 +676,8 @@ export class User {
 			this.driftClient.getPerpMarketAccount(marketIndex),
 			baseAssetAmount,
 			'Initial',
-			this.getUserAccount().maxMarginRatio
+			this.getUserAccount().maxMarginRatio,
+			this.isHighLeverageMode()
 		);
 
 		return freeCollateral.mul(MARGIN_PRECISION).div(new BN(marginRatio));
@@ -1507,7 +1508,8 @@ export class User {
 					market,
 					baseAssetAmount.abs(),
 					marginCategory,
-					this.getUserAccount().maxMarginRatio
+					this.getUserAccount().maxMarginRatio,
+					this.isHighLeverageMode()
 				)
 			);
 
@@ -1972,7 +1974,8 @@ export class User {
 			market,
 			maxSize,
 			marginCategory,
-			this.getUserAccount().maxMarginRatio
+			this.getUserAccount().maxMarginRatio,
+			this.isHighLeverageMode()
 		);
 
 		// use more fesible size since imf factor activated
@@ -1993,7 +1996,8 @@ export class User {
 				market,
 				targetSize,
 				marginCategory,
-				this.getUserAccount().maxMarginRatio
+				this.getUserAccount().maxMarginRatio,
+				this.isHighLeverageMode()
 			);
 			attempts += 1;
 		}
@@ -2152,6 +2156,10 @@ export class User {
 
 	public isBankrupt(): boolean {
 		return (this.getUserAccount().status & UserStatus.BANKRUPT) > 0;
+	}
+
+	public isHighLeverageMode(): boolean {
+		return isVariant(this.getUserAccount().marginMode, 'highLeverage');
 	}
 
 	/**
@@ -2443,7 +2451,9 @@ export class User {
 			const marginRatio = calculateMarketMarginRatio(
 				market,
 				baseAssetAmount.abs(),
-				'Maintenance'
+				'Maintenance',
+				this.getUserAccount().maxMarginRatio,
+				this.isHighLeverageMode()
 			);
 
 			return liabilityValue.mul(new BN(marginRatio)).div(MARGIN_PRECISION);
@@ -2488,7 +2498,8 @@ export class User {
 			market,
 			proposedBaseAssetAmount.abs(),
 			marginCategory,
-			this.getUserAccount().maxMarginRatio
+			this.getUserAccount().maxMarginRatio,
+			this.isHighLeverageMode()
 		);
 		const marginRatioQuotePrecision = new BN(marginRatio)
 			.mul(QUOTE_PRECISION)
@@ -2622,7 +2633,8 @@ export class User {
 			targetMarketIndex,
 			baseSize,
 			collateralIndex,
-			this.getUserAccount().maxMarginRatio
+			this.getUserAccount().maxMarginRatio,
+			false // assume user cant be high leverage if they havent created user account ?
 		);
 	}
 
@@ -3771,7 +3783,8 @@ export class User {
 				perpMarket,
 				worstCaseBaseAmount.abs(),
 				marginCategory,
-				this.getUserAccount().maxMarginRatio
+				this.getUserAccount().maxMarginRatio,
+				this.isHighLeverageMode()
 			)
 		);